+++ packages/client-ts/src/index.ts
@@ -1,11 +1,9 @@
 import Client from './client'
 import Network from './network'
 import Context from './context/context'
-import * as Utils from './utils'
 
 export {
     Client,
     Context,
-    Network,
-    Utils
+    Network
 }