+++ packages/verifiable-credentials/CHANGELOG.md
@@ -1,3 +1,11 @@
+
+- Support proofStrings
+- Support easy creation of a Verida compatible credential record
+- Use new @verida/helpers package
+- Export interfaces
+- Support defining icon of Verida compatible credential record
+- Support auto-generating proofs defined in credential schema
+
 2022-02-09 (v2.0.6)
 -------------------